+++ CHANGELOG.md
@@ -1,5 +1,107 @@
 # Change Log
 
+## [v1.1.1-preview.1](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.1.1-preview.1) (2019-09-24)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/compare/v1.0.8-preview.20...v1.1.1-preview.1)
+
+**Implemented enhancements:**
+
+- Add radio event [\#71](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/pull/71)
+- Add CustomEvent class [\#68](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/pull/68)
+
+**Fixed bugs:**
+
+- Fix unpacking data for custom event [\#69](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/pull/69)
+
+## [v1.0.8-preview.20](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.0.8-preview.20) (2019-07-18)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/compare/v1.0.8-preview.14...v1.0.8-preview.20)
+
+**Documentation and other chores:**
+
+- Bumps nanoFramework.CoreLibrary from 1.2.6-preview.7 to 1.2.6-preview.16 [\#67](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/pull/67)
+
+## [v1.0.8-preview.14](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.0.8-preview.14) (2019-06-23)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/compare/v1.0.8-preview.8...v1.0.8-preview.14)
+
+**Documentation and other chores:**
+
+- Bumps nanoFramework.CoreLibrary from 1.2.6-preview.1 to 1.2.6-preview.7 [\#64](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/pull/64)
+- Work on native assembly declaration [\#62](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/pull/62) [[Breaking-Change](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/labels/Breaking-Change)]
+
+## [v1.0.8-preview.8](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.0.8-preview.8) (2019-06-20)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/compare/v1.0.8-preview.5...v1.0.8-preview.8)
+
+**Documentation and other chores:**
+
+- Bumps nanoFramework.CoreLibrary from 1.2.5-preview.1 to 1.2.6-preview.1 [\#61](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/pull/61)
+
+## [v1.0.8-preview.5](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.0.8-preview.5) (2019-06-12)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/compare/v1.0.8-preview.1...v1.0.8-preview.5)
+
+**Documentation and other chores:**
+
+- Bumps nanoFramework.CoreLibrary from 1.2.4-preview.1 to 1.2.5-preview.1 [\#60](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/pull/60)
+
+## [v1.0.8-preview.1](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.0.8-preview.1) (2019-06-12)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/compare/v1.0.7-preview.5...v1.0.8-preview.1)
+
+## [v1.0.7-preview.5](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.0.7-preview.5) (2019-06-12)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/compare/v1.0.7-preview.63...v1.0.7-preview.5)
+
+**Documentation and other chores:**
+
+- Bumps nanoFramework.CoreLibrary from 1.2.2-preview.20 to 1.2.4-preview.1 [\#59](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/pull/59)
+
+## [v1.0.7-preview.63](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.0.7-preview.63) (2019-06-11)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/compare/v1.0.6-preview-059...v1.0.7-preview.63)
+
+**Documentation and other chores:**
+
+- Bumps nanoFramework.CoreLibrary from 1.2.1-preview-013 to 1.2.2-preview.20 [\#58](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/pull/58)
+
+## [v1.0.6-preview-059](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.0.6-preview-059) (2019-06-11)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/compare/v1.0.6-preview-001...v1.0.6-preview-059)
+
+**Documentation and other chores:**
+
+- Bumps nanoFramework.CoreLibrary from 1.2.1-preview-001 to 1.2.1-preview-013 [\#57](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/pull/57)
+
+## [v1.0.6-preview-001](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.0.6-preview-001) (2019-04-24)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/compare/v1.0.5-preview-019...v1.0.6-preview-001)
+
+## [v1.0.5-preview-019](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.0.5-preview-019) (2019-04-23)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/compare/v1.0.5-preview-012...v1.0.5-preview-019)
+
+**Documentation and other chores:**
+
+- Bumps nanoFramework.CoreLibrary from 1.2.0 to 1.2.1-preview-001 [\#55](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/pull/55)
+- Bumps nanoFramework.CoreLibrary from 1.1.1 to 1.2.0 [\#54](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/pull/54)
+
+## [v1.0.5-preview-012](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.0.5-preview-012) (2019-04-04)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/compare/v1.0.5-preview-010...v1.0.5-preview-012)
+
+## [v1.0.5-preview-010](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.0.5-preview-010) (2019-04-03)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/compare/v1.0.5-preview-008...v1.0.5-preview-010)
+
+**Fixed bugs:**
+
+- Add missing value to Storage event category [\#53](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/pull/53)
+
+## [v1.0.5-preview-008](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.0.5-preview-008) (2019-03-10)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/compare/v1.0.5-preview-007...v1.0.5-preview-008)
+
+## [v1.0.5-preview-007](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.0.5-preview-007) (2019-03-10)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/compare/v1.0.5-preview-005...v1.0.5-preview-007)
+
+## [v1.0.5-preview-005](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.0.5-preview-005) (2019-03-10)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/compare/v1.0.5-preview-003...v1.0.5-preview-005)
+
+## [v1.0.5-preview-003](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.0.5-preview-003) (2019-03-06)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/compare/v1.0.4...v1.0.5-preview-003)
+
+**Implemented enhancements:**
+
+- Add Storage event [\#52](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/pull/52)
+
 ## [v1.0.4](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.0.4) (2019-01-21)
 [Full Changelog](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/compare/v1.0.2...v1.0.4)
 
@@ -29,6 +131,7 @@
 
 **Documentation and other chores:**
 
+- Update readme.txt [\#66](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/pull/66)
 - Bumps nanoFramework.CoreLibrary from 1.0.6-preview-010 to 1.0.6-preview-021 [\#45](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/pull/45)
 
 ## [v1.0.2-preview-017](https://github.com/nanoframework/lib-nanoFramework.Runtime.Events/tree/v1.0.2-preview-017) (2018-11-20)